update jenkins cli

+++ jenkins/cli/create-credentials-domain-with-xml.sh
@@ -1,23 +1,36 @@
 #!/bin/bash
 
-source ./env.sh
+source ./env-base.sh
+source ./env-credentials.sh
 
-CREDENTIAL_DOMAIN="cicd"
+CREDENTIALS_DOMAIN="${CREDENTIALS_DOMAIN}"
 
+######################################################################
+
+if [ ! -f credentials-domain.xml ]; then
 cat <<EOF > ./credentials-domain.xml
 <com.cloudbees.plugins.credentials.domains.Domain plugin="credentials@2.3.12">
-    <name>${CREDENTIAL_DOMAIN}</name>
+    <name>${CREDENTIALS_DOMAIN}</name>
     <specifications/>
 </com.cloudbees.plugins.credentials.domains.Domain>
 EOF
+fi
 
-${JAVA_HOME}/bin/java -jar ${JENKINS_FILE_DIR}/jenkins-cli.jar -s ${JENKINS_URL} -webSocket -auth ${JENKINS_USER}:${JENKINS_API_TOKEN} \
-    create-credentials-domain-with-xml "SystemCredentialsProvider::SystemContextResolver::jenkins" < credentials-domain.xml
-# ${JAVA_HOME}/bin/java -jar ${JENKINS_FILE_DIR}/jenkins-cli.jar -s ${JENKINS_URL} -webSocket -auth ${JENKINS_USER}:${JENKINS_API_TOKEN} \
-#     create-credentials-domain-with-xml "system::system::jenkins" < credentials-domain.xml
+######################################################################
 
-### get-credentials-domain-as-xml
-# ${JAVA_HOME}/bin/java -jar ${JENKINS_FILE_DIR}/jenkins-cli.jar -s ${JENKINS_URL} -webSocket -auth ${JENKINS_USER}:${JENKINS_API_TOKEN} \
-#     get-credentials-domain-as-xml "SystemCredentialsProvider::SystemContextResolver::jenkins" "${CREDENTIAL_DOMAIN}"
-# ${JAVA_HOME}/bin/java -jar ${JENKINS_FILE_DIR}/jenkins-cli.jar -s ${JENKINS_URL} -webSocket -auth ${JENKINS_USER}:${JENKINS_API_TOKEN} \
-#     get-credentials-domain-as-xml "system::system::jenkins" "${CREDENTIAL_DOMAIN}"
+function create_credentials_domain_with_xml {
+    ${JAVA_HOME}/bin/java -jar ${JENKINS_FILE_DIR}/jenkins-cli.jar \
+        -s ${JENKINS_URL} \
+        -webSocket \
+        -auth ${JENKINS_USER}:${JENKINS_API_TOKEN} \
+        create-credentials-domain-with-xml "SystemCredentialsProvider::SystemContextResolver::jenkins" < credentials-domain.xml
+    # ${JAVA_HOME}/bin/java -jar ${JENKINS_FILE_DIR}/jenkins-cli.jar \
+    #     -s ${JENKINS_URL} \
+    #     -webSocket \
+    #     -auth ${JENKINS_USER}:${JENKINS_API_TOKEN} \
+    #     create-credentials-domain-with-xml "system::system::jenkins" < credentials-domain.xml
+}
+
+######################################################################
+
+create_credentials_domain_with_xml
