+++ jenkins/cli/create-credentials-with-xml-aws.sh
@@ -1,37 +1,48 @@
 #!/bin/bash
 
-source ./env.sh
+source ./env-base.sh
+source ./env-credentials.sh
 
-CREDENTIAL_DOMAIN="${${CREDENTIAL_DOMAIN}:-cicd}"
-CREDENTIAL_ACCESSKEY="${1:-accesskey}"
-CREDENTIAL_SECRETKEY="${2:-secretkey}"
-CREDENTIAL_DESCRIPTION="${CREDENTIAL_DESCRIPTION}"
+CREDENTIALS_DOMAIN="${CREDENTIALS_DOMAIN}"
+CREDENTIALS_ACCESSKEY="${CREDENTIALS_ACCESSKEY}"
+CREDENTIALS_SECRETKEY="${CREDENTIALS_SECRETKEY}"
+CREDENTIALS_DESCRIPTION="${CREDENTIALS_DESCRIPTION}"
+CREDENTIALS_UUID="$(uuidgen)"
 
-CREDENTIAL_UUID="$(uuidgen)"
-echo "UUID: ${CREDENTIAL_UUID}"
+echo "UUID: ${CREDENTIALS_UUID}"
 
+######################################################################
+
+if [ ! -f credentials-aws.xml ]; then
 cat <<EOF > ./credentials-aws.xml
 <com.cloudbees.jenkins.plugins.awscredentials.AWSCredentialsImpl plugin="aws-credentials@1.28">
     <scope>GLOBAL</scope>
-    <id>${CREDENTIAL_UUID}</id>
-    <description>${CREDENTIAL_DESCRIPTION}</description>
-    <accessKey>${CREDENTIAL_ACCESSKEY}</accessKey>
-    <secretKey>${CREDENTIAL_SECRETKEY}</secretKey>
+    <id>${CREDENTIALS_UUID}</id>
+    <description>${CREDENTIALS_DESCRIPTION}</description>
+    <accessKey>${CREDENTIALS_ACCESSKEY}</accessKey>
+    <secretKey>${CREDENTIALS_SECRETKEY}</secretKey>
     <iamRoleArn></iamRoleArn>
     <iamMfaSerialNumber></iamMfaSerialNumber>
 </com.cloudbees.jenkins.plugins.awscredentials.AWSCredentialsImpl>
 EOF
+fi
+
+######################################################################
+
+function create_credentials_with_xml_aws {
+    ${JAVA_HOME}/bin/java -jar ${JENKINS_FILE_DIR}/jenkins-cli.jar \
+        -s ${JENKINS_URL} \
+        -webSocket \
+        -auth ${JENKINS_USER}:${JENKINS_API_TOKEN} \
+        create-credentials-with-xml "SystemCredentialsProvider::SystemContextResolver::jenkins" "${CREDENTIALS_DOMAIN}" < credentials-aws.xml
+    # ${JAVA_HOME}/bin/java -jar ${JENKINS_FILE_DIR}/jenkins-cli.jar \
+    #     -s ${JENKINS_URL} \
+    #     -webSocket \
+    #     -auth ${JENKINS_USER}:${JENKINS_API_TOKEN} \
+    #     create-credentials-with-xml "system::system::jenkins" "${CREDENTIALS_DOMAIN}" < credentials-aws.xml
 }
 
-${JAVA_HOME}/bin/java -jar ${JENKINS_FILE_DIR}/jenkins-cli.jar -s ${JENKINS_URL} -webSocket -auth ${JENKINS_USER}:${JENKINS_API_TOKEN} \
-    create-credentials-with-xml "SystemCredentialsProvider::SystemContextResolver::jenkins" "${CREDENTIAL_DOMAIN}" < credentials-aws.xml
-# ${JAVA_HOME}/bin/java -jar ${JENKINS_FILE_DIR}/jenkins-cli.jar -s ${JENKINS_URL} -webSocket -auth ${JENKINS_USER}:${JENKINS_API_TOKEN} \
-#     create-credentials-with-xml "system::system::jenkins" "${CREDENTIAL_DOMAIN}" < credentials-aws.xml
+######################################################################
 
-rm credentials-aws.xml
-
-### get-credentials-as-xml
-# ${JAVA_HOME}/bin/java -jar ${JENKINS_FILE_DIR}/jenkins-cli.jar -s ${JENKINS_URL} -webSocket -auth ${JENKINS_USER}:${JENKINS_API_TOKEN} \
-#     get-credentials-as-xml "SystemCredentialsProvider::SystemContextResolver::jenkins" "${CREDENTIAL_DOMAIN}" "${CREDENTIAL_UUID}"
-# ${JAVA_HOME}/bin/java -jar ${JENKINS_FILE_DIR}/jenkins-cli.jar -s ${JENKINS_URL} -webSocket -auth ${JENKINS_USER}:${JENKINS_API_TOKEN} \
-#     get-credentials-as-xml "system::system::jenkins" "${CREDENTIAL_DOMAIN}" "${CREDENTIAL_UUID}"
+create_credentials_with_xml_aws
+# rm -f ./credentials-aws.xml
